+++ ecs/world_examples_test.go
@@ -232,7 +232,7 @@ func ExampleWorld_SetListener() {
 	world.SetListener(listener)
 
 	world.NewEntity()
-	// Output: &{{1 0} {0 0} {0 0} [] [] [] 1}
+	// Output: &{{1 0} {0 0} {0 0} [] [] [] 1 {0 0} {0 0} false}
 }
 
 func ExampleWorld_Stats() {