+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * Re-introduce `World.Batch` for batch-processing of entities (add/remove/exchange) (#264)
 * New method `Builder.Add` for adding components with a target to entities (#264)
 * New method `Batch.SetRelation` for batch-setting entity relations (#265)
+* Sends an `EntityEvent` to the world listener on relation target changes (#265)
 
 ### Performance